requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0

# Diffusion (generation pipeline)
diffusers>=0.31.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0

# Object Detection
ultralytics>=8.1.0

# Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data & Visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
tqdm>=4.65.0

# Web Demo
gradio>=4.0.0

# System
psutil>=5.9.0

# Dev
pytest>=7.4.0
